+++ jxls-site/docs/changes.md
@@ -8,6 +8,8 @@ highlights:
 
 tickets:
 
+- [#396 POI bug 66679 compensation](https://github.com/jxlsteam/jxls/issues/396) activate with withCellStyleGeneralEnsurer(),
+  also new: template preprocessing
 - [#394 POI 5.5.1](https://github.com/jxlsteam/jxls/pull/394)
 - [#377 Add link command](https://github.com/jxlsteam/jxls/pull/377) *- contribution by [kingwtd](https://github.com/kingwtd)*
 - [#375 Same multisheet sheet name and template sheet name results in removed sheet without any error](https://github.com/jxlsteam/jxls/issues/375)